[Monitor][Query] Extract enums to separate file
Enums are moved from the models file to a separate file
to better organize and avoid a circular dependency when trying
to import LogsQueryError from _exceptions.py for typing purposes.
